# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Network Atlas API",
    "description": "MVP-1 visualization-first contract: Linux SSH evidence, deterministic local projection, optional Agent assistance, and a single-owner remote release boundary",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/model-provider": {
      "get": {
        "tags": [
          "m3"
        ],
        "operationId": "get_model_provider",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProviderResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "owner_session": []
          }
        ]
      },
      "put": {
        "tags": [
          "m3"
        ],
        "operationId": "put_model_provider",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Stable key for safe request retries",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelProviderPutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProviderResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "owner_session": []
          }
        ]
      }
    },
    "/api/v1/model-provider/test": {
      "post": {
        "tags": [
          "m3"
        ],
        "operationId": "test_model_provider",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Stable key for safe request retries",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProviderTestResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "owner_session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message",
          "details",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {},
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        }
      },
      "ApiMeta": {
        "type": "object",
        "required": [
          "request_id",
          "revision",
          "generated_at",
          "freshness",
          "data_source"
        ],
        "properties": {
          "data_source": {
            "$ref": "#/components/schemas/DataSourceDescriptor"
          },
          "freshness": {
            "$ref": "#/components/schemas/Freshness"
          },
          "generated_at": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "revision": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "DataSourceDescriptor": {
        "type": "object",
        "required": [
          "kind",
          "status",
          "label"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/DataSourceKind"
          },
          "label": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DataSourceStatus"
          }
        }
      },
      "DataSourceKind": {
        "type": "string",
        "enum": [
          "fixture",
          "real"
        ]
      },
      "DataSourceStatus": {
        "type": "string",
        "enum": [
          "fresh",
          "stale",
          "unavailable"
        ]
      },
...
      "Freshness": {
        "type": "string",
        "enum": [
          "fresh",
          "stale",
          "unavailable"
        ]
      },
...
      "ModelProviderData": {
        "type": "object",
        "required": [
          "base_url",
          "model",
          "key_state",
          "revision",
          "updated_at"
        ],
        "properties": {
          "base_url": {
            "type": "string"
          },
          "key_state": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "revision": {
            "type": "integer",
            "format": "int64"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "ModelProviderPutRequest": {
        "type": "object",
        "required": [
          "base_url",
          "model",
          "credential_ref"
        ],
        "properties": {
          "base_url": {
            "type": "string"
          },
          "credential_ref": {
            "type": "string"
          },
          "model": {
            "type": "string"
          }
        }
      },
      "ModelProviderResponse": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ModelProviderData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "ModelProviderTestData": {
        "type": "object",
        "required": [
          "test_id",
          "state",
          "latency_ms"
        ],
        "properties": {
          "error_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "latency_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "state": {
            "$ref": "#/components/schemas/ModelProviderTestState"
          },
          "test_id": {
            "type": "string"
          }
        }
      },
      "ModelProviderTestResponse": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ModelProviderTestData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "ModelProviderTestState": {
        "type": "string",
        "enum": [
          "reachable",
          "failed"
        ]
      },
...
    },
    "securitySchemes": {
      "owner_session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "network_atlas_session",
        "description": "HttpOnly, Secure, SameSite=Strict single-owner session cookie"
      }
    }
  },
...
}